// src/api/generated/model/createCategoryDto.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * Personal Blog API
 * 个人博客 API 文档
 * OpenAPI spec version: 1.0
 */
import type { CreateCategoryDtoColor } from './createCategoryDtoColor';

export interface CreateCategoryDto {
  /**
   * 唯一标识（仅字母/数字/下划线，<=12，强制小写）
   * @maxLength 12
   */
  slug: string;
  /** @maxLength 50 */
  name: string;
  description?: string;
  /** 必须从主题色中选择 */
  color: CreateCategoryDtoColor;
  /**
   * 排序权重（越小越靠前）
   * @minimum 0
   */
  sortOrder?: number;
  /** 是否启用 */
  isActive?: boolean;
}