// src/api/generated/model/updateCategoryDto.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * Personal Blog API
 * 个人博客 API 文档
 * OpenAPI spec version: 1.0
 */
import type { UpdateCategoryDtoDescription } from './updateCategoryDtoDescription';
import type { UpdateCategoryDtoColor } from './updateCategoryDtoColor';

export interface UpdateCategoryDto {
  /** @maxLength 50 */
  name?: string;
  description?: UpdateCategoryDtoDescription;
  /** 必须从主题色中选择 */
  color?: UpdateCategoryDtoColor;
  /**
   * 排序权重（越小越靠前）
   * @minimum 0
   */
  sortOrder?: number;
  /** 是否启用 */
  isActive?: boolean;
}